Major update
The Google Tasks API library was dropped in favor of an implementation that uses Retrofit to connect to the Tasks API. Essentially, the first time the user signs in, an access token is fetched using GoogleAuthUtil. This access token is then used for every call to the Tasks API. When the access token expires, a new one is requested without interrupting the flow of the app. The GET_ACCOUNTS permission was therefore no longer required, and because it was the only "dangerous" permission, Permiso was removed as well. This, along with other improvements, allowed for a smaller app size.

The full list of changes:

- TargetSDK 24 :)
- Add Task Detail activity. This allows the user to view more details about a specific task, edit it, delete it or mark it as completed.
- Add Edit Task activity. This is where the user can modify the task's fields, such as title or due date.
- Add shared element and content transitions for the Task Detail activity (on Lollipop and above)
- Add data binding for Task Detail and Edit Task activities
- Add 3 custom BindingAdapters
- Add basic handling of user recoverable errors related to the Tasks API (WIP)
- Add caching of user picture and cover so that they are shown even offline
- Improve empty view for task lists
- Improve task & task list refreshing by caching the response ETag and providing it on every call so that refreshing happens only when data was modified
- Improve ProGuard rules
- Improve app styles
- Restructure packages, Dagger components & modules
- Update libraries
- Many other fixes and improvements