Add basic reminder functionality
The reminder information could not be stored in the Task class because it would be overridden at the next refresh. This is an issue with Realm, fields that do not exist in the JSON response get initialized to their default value and that's how the reminder is lost. A new TTask class was added, which contains a Task and any additional information not provided in the response.
- Add basic notification for reminders
- Fix due date resetting, now takes into account the reminder
- Fix task loading after the Realm schema has changed
- Tweaks to the task item layout
- Fix reminder icon not showing
- Remove hungarian notation since it just makes the code more unreadable and is useless in an IDE such as Android Studio
- Make Utils classes un-instantiable
- Refactoring