Kotlin conversion
The Kotlin version is finally up! There were a bunch of breaking changes, mostly because of the different generics system in Kotlin. As a result, FastAdapter was replaced by FlexibleAdapter because it plays much better with Kotlin. Libraries were updated to their most recent versions as of the commit date. The app shouldn't crash at any point in the current state but some minor things don't work as expected, eg the arrow next to the task categories.
The Java version is now deprecated, <3 Kotlin